core: Store condition block results in plan
In order to include condition block results in the JSON plan output, we
must store them in the plan and its serialization. This commit achives
this with a simple plans.Conditions object and a sync wrapper.

Terraform can evaluate condition blocks multiple times, so we must be
able to update the result for each one. To achieve this, the Conditions
object is a map with keys representing the condition block's address.
Condition blocks are not referenceable in any other context, so this
address cannot be used anywhere in the configuration.

The commit includes a new test case for the JSON output of a
refresh-only plan, which is currently the only way for a failing
condition result to be rendered through this path.